move json types to separate file
